chnage color, fix some stuff
