+++ packages/web/components/Player/PlayerProfilePicture.tsx
@@ -2,8 +2,8 @@ import type { AvatarProps } from '@metafam/ds';
 import { SquareImage } from 'components/SquareImage';
 import type { Player } from 'graphql/autogen/types';
 import type { GuildPlayer } from 'graphql/types';
-import React, { useMemo, useState, useEffect } from 'react';
-import { getPlayerImage } from 'utils/playerHelpers';
+import React from 'react';
+import { useProfileImageOnload } from 'lib/hooks/useProfileImageOnload';
 
 type PlayerProfilePictureProps = AvatarProps & {
   player?: Player | GuildPlayer;
@@ -14,33 +14,11 @@ export const PlayerProfilePicture: React.FC<PlayerProfilePictureProps> = ({
   src,
   ...props
 }) => {
-  const avatarImg = useMemo(() => getPlayerImage(player), [player, src]);
-  
-  const [imageURL, setImageURL] = useState(avatarImg.profileIcon);
-  const profileImage = avatarImg.link;
-
-  useEffect(() => {
-    if (!profileImage) return;
-    const image = new Image();
-    image.src = profileImage;
-
-    image.onload = () => {
-      setImageURL(profileImage);
-    };
-
-    image.onerror = () => {
-      console.error('Original image failed to load.');
-    };
-
-    return () => {
-      image.onload = null;
-      image.onerror = null;
-    };
-  }, [profileImage]);
+  const imageUrl = useProfileImageOnload({ player });
 
   return (
     <>
-      <SquareImage src={imageURL} {...props} />
+      <SquareImage src={imageUrl as string ?? undefined} {...props} />
     </>
   );
 };
