+++ packages/web/components/Player/PlayerProfilePictureRound.tsx
@@ -2,8 +2,8 @@ import type { AvatarProps } from '@metafam/ds';
 import { RoundImage } from 'components/RoundImage';
 import type { Player } from 'graphql/autogen/types';
 import type { GuildPlayer } from 'graphql/types';
-import React, { useMemo } from 'react';
-import { getPlayerImage } from 'utils/playerHelpers';
+import { useProfileImageOnload } from 'lib/hooks/useProfileImageOnload';
+import React from 'react';
 
 type PlayerProfilePictureProps = AvatarProps & {
   player?: Player | GuildPlayer;
@@ -14,11 +14,11 @@ export const PlayerProfilePictureRound: React.FC<PlayerProfilePictureProps> = ({
   src,
   ...props
 }) => {
-  const avatarImg = useMemo(() => src ?? getPlayerImage(player), [player, src]);
+  const imageUrl = useProfileImageOnload({ player });
 
   return (
     <>
-      <RoundImage src={avatarImg ?? undefined} {...props} />
+      <RoundImage src={imageUrl as string ?? undefined} {...props} />
     </>
   );
 };
